Remove `DefaultMovement` implementation, refactor unit-to-army data structures and counts, and extend protobuf definitions with new message types and fields.

+++ server/internal/server/objects/units/Attack/IAttack.go
@@ -3,9 +3,9 @@ package Attack
 import "server/internal/server"
 
 type IAttack interface {
-	UnitId() uint64
+	UnitId() int32
 	ClientId() uint64
-	InitAttack(UnitId uint64, client *server.Client, enemyClient *server.Client)
+	InitAttack(UnitId int32, client *server.Client, enemyClient *server.Client)
 	Attack(unitId uint64)
 	SendAttackPacket(unitId uint64)
 }
